xml update
fixed the  ones that were incorrectly given the n=1 tag and got ride of n=1 as there isn't a 6 game currently
also I added comments to the scheme so people would know what the n= tags meant.
Additionally, I fixed the tiers of the characters that were recently added to our xml and oved the 4 DLC characters whose tiers I was unsure of to the bottom of the xml.